refactor(2343): replace subscribe activity modal with confirmation modal
