+++ e2e/tests/student/lifeProject/activitiesCatalog/activitiesCatalog.feature
@@ -48,7 +48,7 @@ Feature: Student Project Activities Catalog Page
       And the activity preview is correctly displayed
       And the activity preview subscribe button is visible
 
-  Rule: Activity subscription - Desktop
+Rule: Activity subscription - Desktop
 
     Background:
       When the user clicks on the activity preview subscribe button
@@ -58,24 +58,9 @@ Feature: Student Project Activities Catalog Page
         Then the subscribe activity modal is displayed
 
       @medium @subscribe
-      Scenario: Student can see the cancel subscribe modal
-        When the user fills in the subscribe activity modal period start date
-        And the user clicks on the activity preview subscribe modal cancel button
-        Then the cancel subscribe activity confirm modal is displayed
-
-      @low @subscribe
-      Scenario: Student can confirm the cancel subscribe modal
-        When the user fills in the subscribe activity modal period start date
-        And the user clicks on the activity preview subscribe modal cancel button and confirms
-        Then the cancel subscribe activity confirm modal is hidden
-        And the subscribe activity modal is hidden
-
-      @low @subscribe
-      Scenario: Student can cancel the cancel subscribe modal
-        When the user fills in the subscribe activity modal period start date
-        And the user clicks on the activity preview subscribe modal cancel button and cancels
-        Then the cancel subscribe activity confirm modal is hidden
-        And the subscribe activity modal is displayed
+      Scenario: Student can cancel the subscribe activity modal
+        When the user clicks on the activity preview subscribe modal cancel button
+        Then the subscribe activity modal is hidden
 
       @high @subscribe @skip
       Scenario: Student can subscribe to the activity and unsubscribe from it right after
