+++ src/features/student/buildProject/locales/en.json
@@ -39,6 +39,11 @@
           }
         },
         "overlays": {
+          "SubscribeActivityConfirmModal": {
+            "error": "An error occurred while subscribing to the activity. Please try again later.",
+            "success": "You have successfully subscribed. Find this activity in your activity library.",
+            "title": "Are you sure you want to subscribe to the activity: {title}?"
+          },
           "UnsubscribeActivitiesConfirmModal": {
             "description": "This action will result in the permanent loss of all associated data and actions.",
             "error": "An error occurred while unsubscribing from the activity. Please try again later. | An error occurred while unsubscribing from the activities. Please try again later.",
@@ -76,16 +81,6 @@
           "ProjectActivitiesCatalogView": {
             "buttons": {
               "access": "Access my activity"
-            },
-            "overlays": {
-              "ActivitySubscribeModal": {
-                "error": "An error occurred while subscribing to the activity. Please try again later.",
-                "success": "You have successfully subscribed. Find this activity in your activity library.",
-                "title": "Subscribe to the activity {title}"
-              },
-              "CancelSubscribeActivityModal": {
-                "title": "Are you sure you want to cancel your subscription to the activity?"
-              }
             },
             "title": "All available activities",
             "unsubscribe": "Unsubscribe"
