--- src/features/student/buildProject/views/ProjectActivitiesCatalogView/components/overlays/CancelSubscribeActivityConfirmModal/CancelSubscribeActivityConfirmModal.stub.ts
@@ -1,13 +1,0 @@
-export const CancelSubscribeActivityConfirmModalStub = defineComponent({
-  name: 'CancelSubscribeActivityConfirmModal',
-  props: {
-    show: {
-      type: Boolean,
-      required: true
-    }
-  },
-  emits: ['cancel', 'confirm'],
-  template: `<div v-if="show">
-    <p>CancelSubscribeActivityConfirmModalStub</p>
-  </div>`
-})
